Add contract price to get price

+++ src/Product/Postex.Product.Application/Dtos/Contratcs/ContractBoxPriceDto.cs
@@ -1,9 +1,4 @@
 ﻿using Postex.Product.Domain.Contracts;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Postex.Product.Application.Dtos.Contratcs
 {
@@ -20,8 +15,8 @@ namespace Postex.Product.Application.Dtos.Contratcs
         public int? ProvinceId { get; set; }
         public int? CityId { get; set; }
         public int? CustomerId { get; set; }
-        public double SalePrice { get; set; }
-        public double BuyPrice { get; set; }
+        public decimal SalePrice { get; set; }
+        public decimal BuyPrice { get; set; }
         public string Description { get; set; }
         public string LevelPrice { get; set; }
         public bool IsActive { get; set; }
