+++ src/Product/Postex.Product.Application/Features/Contratcs/ContractBoxPrices/Commands/Update/UpdateContractBoxPriceCommand.cs
@@ -1,10 +1,5 @@
 ﻿using Postex.Product.Application.Contracts;
 using Postex.Product.Application.Dtos.Contratcs;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Postex.Product.Application.Features.Contratcs.ContractBoxPrices.Commands.Update
 {
@@ -15,8 +10,8 @@ namespace Postex.Product.Application.Features.Contratcs.ContractBoxPrices.Comman
         public int? ProvinceId { get; set; }
         public int? CityId { get; set; }
         public int? CustomerId { get; set; }
-        public double SalePrice { get; set; }
-        public double BuyPrice { get; set; }
+        public decimal SalePrice { get; set; }
+        public decimal BuyPrice { get; set; }
         public string Description { get; set; }
         public bool IsActive { get; set; }
     }
