+++ src/Product/Postex.Product.Application/Features/Contratcs/ContractValueAddeds/Commands/Update/UpdateContractValueAddedCommand.cs
@@ -11,8 +11,8 @@ namespace Postex.Product.Application.Features.Contratcs.ContractValueAddeds.Comm
         public int? ProvinceId { get; set; }
         public int? CityId { get; set; }
         public bool IsActive { get; set; }
-        public double SalePrice { get; set; }
-        public double BuyPrice { get; set; }
+        public decimal SalePrice { get; set; }
+        public decimal BuyPrice { get; set; }
         public string Description { get; set; }
     }
 }
